requirements.txt
streamlit
numpy
plotly
scipy
pandas
Pillow